Moved the exercise into a module
