+++ src/windows/workbook/mod.rs
@@ -1,3 +1,4 @@
 mod book;
+mod exercise;
 
 pub use book::BookWindow as WorkbookWindow;
